rag-chatbot-backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-core
langchain-groq
sentence-transformers
chromadb
python-multipart
PyMuPDF
pytesseract
pillow
requests
python-dotenv
tqdm
transformers
pymongo
python-dotenv
